feat(MetricCollector): Refactors to make current collector a MetricCollector
refs #48
