+++ internal/collector/collector.go
@@ -2,357 +2,12 @@ package collector
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"github.com/google/uuid"
-	"github.com/turbolytics/collector/internal/collector/state"
-	"github.com/turbolytics/collector/internal/config"
-	"github.com/turbolytics/collector/internal/metrics"
-	"github.com/turbolytics/collector/internal/obs"
-	"github.com/turbolytics/collector/internal/timeseries"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/metric"
-	"go.uber.org/zap"
 	"time"
 )
-
-var meter = otel.Meter("signals-collector")
 
 type Collector interface {
 	Close() error
 	InvokeHandleError(context.Context)
 	Interval() *time.Duration
 	Cron() *string
-}
-
-type MetricCollector struct {
-	Config *config.Config
-
-	logger *zap.Logger
-	now    func() time.Time
-}
-
-func (c *MetricCollector) Close() error {
-	for _, s := range c.Config.Sinks {
-		s.Sinker.Close()
-	}
-	return nil
-}
-
-func (c *MetricCollector) Interval() *time.Duration {
-	return c.Config.Schedule.Interval
-}
-
-func (c *MetricCollector) Cron() *string {
-	return c.Config.Schedule.Cron
-}
-
-func (c *MetricCollector) Transform(ms []*metrics.Metric) error {
-	for _, m := range ms {
-		m.Name = c.Config.Metric.Name
-		m.Type = c.Config.Metric.Type
-
-		// enrich with tags
-		// should these be copied?
-		for _, t := range c.Config.Metric.Tags {
-			m.Tags[t.Key] = t.Value
-		}
-
-	}
-	return nil
-}
-
-func (c *MetricCollector) Source(ctx context.Context) (ms []*metrics.Metric, err error) {
-	start := time.Now().UTC()
-
-	histogram, _ := meter.Float64Histogram(
-		"collector.source.duration",
-		metric.WithUnit("s"),
-	)
-
-	defer func() {
-		duration := time.Since(start)
-
-		histogram.Record(ctx, duration.Seconds(), metric.WithAttributeSet(
-			attribute.NewSet(
-				attribute.String("result.status_code", obs.ErrToStatus(err)),
-				attribute.String("source.type", string(c.Config.Source.Type)),
-			),
-		))
-
-		meter.Int64ObservableGauge(
-			"collector.source.metrics.total",
-			metric.WithInt64Callback(func(_ context.Context, o metric.Int64Observer) error {
-				o.Observe(int64(len(ms)), metric.WithAttributeSet(
-					attribute.NewSet(
-						attribute.String("source.type", string(c.Config.Source.Type)),
-					),
-				))
-				return nil
-			}),
-		)
-
-	}()
-
-	ms, err = c.Config.Source.Sourcer.Source(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return ms, nil
-}
-
-func (c *MetricCollector) Sink(ctx context.Context, metrics []*metrics.Metric) error {
-
-	histogram, _ := meter.Float64Histogram(
-		"collector.sink.duration",
-		metric.WithUnit("s"),
-	)
-	// add tags from config
-
-	// need to add a serializer
-	for _, m := range metrics {
-		bs, err := json.Marshal(m)
-		if err != nil {
-			return err
-		}
-		for _, s := range c.Config.Sinks {
-			start := time.Now().UTC()
-
-			_, err := s.Sinker.Write(bs)
-
-			duration := time.Since(start)
-			histogram.Record(ctx, duration.Seconds(), metric.WithAttributeSet(
-				attribute.NewSet(
-					attribute.String("result.status_code", obs.ErrToStatus(err)),
-					attribute.String("sink.name", string(s.Type)),
-				),
-			))
-
-			if err != nil {
-				return err
-			}
-
-		}
-	}
-	return nil
-}
-
-// InvokeHandleError will log any Invoke errors and not return them.
-// Useful for async scheduling.
-func (c *MetricCollector) InvokeHandleError(ctx context.Context) {
-	_, err := c.Invoke(ctx)
-	if err != nil {
-		c.logger.Error(err.Error())
-	}
-}
-
-func (c *MetricCollector) invokeTick(ctx context.Context, id uuid.UUID) ([]*metrics.Metric, error) {
-	ms, err := c.Source(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	// only sink if metrics are present:
-	if len(ms) == 0 {
-		c.logger.Warn(
-			"collector.Invoke",
-			zap.String("msg", "no metrics found"),
-			zap.String("id", id.String()),
-			zap.String("name", c.Config.Name),
-		)
-		return ms, err
-	}
-
-	if err = c.Transform(ms); err != nil {
-		return ms, err
-	}
-
-	if err = c.Sink(ctx, ms); err != nil {
-		return ms, err
-	}
-
-	return ms, err
-}
-
-func (c *MetricCollector) invokeWindowSourceAndSave(ctx context.Context, id uuid.UUID, window timeseries.Bucket) ([]*metrics.Metric, error) {
-	c.logger.Info(
-		"collector.invokeWindow",
-		zap.String("msg", "invoking for window"),
-		zap.String("window.start", window.Start.String()),
-		zap.String("window.end", window.End.String()),
-		zap.String("id", id.String()),
-		zap.String("name", c.Config.Name),
-	)
-
-	// it is passed the window, collect data for the window
-	// get start of window and end of window
-	ctx = context.WithValue(ctx, "window.start", window.Start)
-	ctx = context.WithValue(ctx, "window.end", window.End)
-	ms, err := c.Source(ctx)
-	if err != nil {
-		return ms, err
-	}
-
-	err = c.Config.StateStore.Storer.SaveInvocation(&state.Invocation{
-		CollectorName: c.Config.Name,
-		Time:          c.now(),
-		Window:        &window,
-	})
-	return ms, err
-}
-
-// invokeWindow uses the state store to check if a full window has elapsed.
-// invokeWindow will only source data when a full window has elapsed.
-// TODO - What happens when a window is changed in the config?
-func (c *MetricCollector) invokeWindow(ctx context.Context, id uuid.UUID) ([]*metrics.Metric, error) {
-	var ms []*metrics.Metric
-	var err error
-	// TODO invokeWindow should handle gaps in windows.
-	i, err := c.Config.StateStore.Storer.MostRecentInvocation(
-		ctx,
-		c.Config.Name,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	// no previous invocations exist, just perform the current collection
-	// and save.
-	if i == nil {
-		// get the current completed window
-		window := timeseries.LastCompleteBucket(
-			c.now(),
-			*(c.Config.Source.Sourcer.Window()),
-		)
-		ms, err = c.invokeWindowSourceAndSave(ctx, id, window)
-	} else {
-		// a previous invocation exists.
-		// Get all buckets that have passed since invocation end
-
-		// truncating the current time to the duration....
-		windows, err := timeseries.TimeBuckets(
-			i.Window.End,
-			c.now(),
-			*(c.Config.Source.Sourcer.Window()),
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		// if no window has passed return and wait
-		if len(windows) == 0 {
-			return nil, nil
-		} else if len(windows) > 1 {
-			c.logger.Error(
-				"collector.invokeWindow",
-				zap.String("msg", "multiple windows detected"),
-				zap.Int("windows", len(windows)),
-				zap.String("id", id.String()),
-				zap.String("name", c.Config.Name),
-			)
-			return nil, fmt.Errorf("backfilling multiple windows not yet supported: %v", windows)
-		}
-		window := windows[0]
-		ms, err = c.invokeWindowSourceAndSave(ctx, id, window)
-	}
-
-	if err = c.Transform(ms); err != nil {
-		return ms, err
-	}
-
-	if err = c.Sink(ctx, ms); err != nil {
-		return ms, err
-	}
-
-	return ms, err
-}
-
-func (c *MetricCollector) Invoke(ctx context.Context) (ms []*metrics.Metric, err error) {
-	start := time.Now().UTC()
-
-	histogram, _ := meter.Float64Histogram(
-		"collector.invoke.duration",
-		metric.WithUnit("s"),
-	)
-
-	counter, _ := meter.Int64Counter(
-		"collector.invoke.count",
-	)
-
-	defer func() {
-		duration := time.Since(start)
-
-		counter.Add(ctx, 1, metric.WithAttributeSet(
-			attribute.NewSet(
-				attribute.String("result.status_code", obs.ErrToStatus(err)),
-				attribute.String("collector.name", c.Config.Name),
-			),
-		))
-
-		histogram.Record(ctx, duration.Seconds(), metric.WithAttributeSet(
-			attribute.NewSet(
-				attribute.String("result.status_code", obs.ErrToStatus(err)),
-				attribute.String("collector.name", c.Config.Name),
-			),
-		))
-	}()
-
-	id := uuid.New()
-	c.logger.Info(
-		"collector.Invoke",
-		zap.String("id", id.String()),
-		zap.String("name", c.Config.Name),
-	)
-	ctx = context.WithValue(ctx, "id", id)
-
-	// MetricCollector supports multiple sourcing strategies.
-	// The simplest is "tick" strategy which just invokes
-	// the sourcer without any additional state necessary
-	// The windowing strategy may result in multiple source
-	// invocations for each window that needs to be executed.
-	switch c.Config.Source.Strategy {
-	case config.TypeSourceStrategyTick:
-		ms, err = c.invokeTick(ctx, id)
-	case config.TypeSourceStrategyWindow:
-		ms, err = c.invokeWindow(ctx, id)
-	default:
-		return nil, fmt.Errorf("strategy: %q not supported", c.Config.Source.Strategy)
-	}
-
-	return ms, err
-}
-
-type Option func(*MetricCollector)
-
-func WithLogger(l *zap.Logger) Option {
-	return func(c *MetricCollector) {
-		c.logger = l
-	}
-}
-
-func New(config *config.Config, opts ...Option) (*MetricCollector, error) {
-	c := &MetricCollector{
-		Config: config,
-		now: func() time.Time {
-			return time.Now().UTC()
-		},
-	}
-	for _, opt := range opts {
-		opt(c)
-	}
-	return c, nil
-}
-
-func NewFromConfigs(configs []*config.Config, opts ...Option) ([]Collector, error) {
-	var cs []Collector
-	for _, conf := range configs {
-		coll, err := New(conf, opts...)
-		if err != nil {
-			return nil, err
-		}
-		cs = append(cs, coll)
-	}
-	return cs, nil
 }
